Moved components into own folder
